Refactor: Table: exporting utils into own file
